+++ packages/tautulli/lib/src/commands/activity/get_activity.dart
@@ -1,0 +1,9 @@
+part of tautulli;
+
+Future<TautulliActivity> _commandGetActivity(Dio client, {
+    int sessionKey,
+    String sessionId,
+}) async {
+    if(sessionKey != null) assert(sessionId == null, 'sessionKey and sessionId both cannot be defined.');
+    return TautulliActivity();
+}