+++ packages/tautulli/CHANGELOG.md
@@ -1,2 +1,5 @@
+# 0.0.1-pre.2
+- Implemented HTTP client
+- Implemented `arnold()`
 # 0.0.1-pre.1
 - Initial Release